requirements.txt
requests==2.31.0
PyQt6==6.10.2
python-dotenv==1.0.0
